Remove manual `Ember.meta` "slotting".
Ember 2.x _sometimes_ added a property to the object itself (later versions of Ember 2.x series used `WeakMap` when available to avoid this). As of Ember 3.x Ember no longer supports the fallback path for `Ember.meta` which would add a property to the underlying object (and therefore cause shaping issues). 

_tldr;_ this code is no longer needed :wink: